+++ src/app/product/[id]/ui/Option.module.scss
@@ -21,12 +21,6 @@
     width: 100%;
   }
 
-  &__quantity {
-    @include mixins.flex-center;
-    gap: 1.25rem;
-    width: 100%;
-  }
-
   &__button {
     @include button-base;
