+++ src/app/product/[id]/ui/Option.tsx
@@ -1,5 +1,4 @@
-import { SIZE_INVENTORY } from '@/app/lib/constants';
-import { Button } from '@/components';
+import { QuantityController } from '@/components';
 import type { OptionProps } from '@/types/product';
 import { CircleX } from 'lucide-react';
 import { memo } from 'react';
@@ -9,32 +8,13 @@ const Option = ({
   size,
   quantity,
   price,
-  onIncrementButtonClick,
-  onDecrementButtonClick,
   onDeleteButtonClick,
+  ...props
 }: OptionProps) => {
-  const maxStock = (selectedSize: number) =>
-    SIZE_INVENTORY.find(({ size }) => size === selectedSize)?.stock;
-
   return (
     <li key={size} className={styles.option}>
       <p className={styles.option__size}>{size}</p>
-      <div className={styles.option__quantity}>
-        <Button
-          onClick={() => onDecrementButtonClick(size)}
-          disabled={!quantity}
-          text="-"
-          variant="lined--small"
-        />
-
-        <p>{quantity}</p>
-        <Button
-          onClick={() => onIncrementButtonClick(size)}
-          disabled={quantity === maxStock(size)}
-          text="+"
-          variant="lined--small"
-        />
-      </div>
+      <QuantityController {...props} size={size} quantity={quantity} />
 
       <div className={styles.option__price_box}>
         <p className={styles.option__price}>
