+++ src/components/index.ts
@@ -2,3 +2,4 @@ export { default as Card } from './Card';
 export { default as Button } from './Button';
 export { default as CartBadge } from './CartBadge';
 export { default as Image } from './Image';
+export { default as QuantityController } from './QuantityController';
